pyproject.toml
[project]
name = "sunglass-store-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.10.1",
    "dbt-postgres>=1.9.0",
    "dlt[duckdb,filesystem,postgres]>=1.12.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.45.1",
    "tqdm>=4.67.1",
    "watchdog>=6.0.0",
]
